Support for Orange Pi Zero 3 in mainline Linux
For things I have tested and are functional out of the box in kernel 6.5.3:
-Ethernet
-SD card
-USB

WiFi is not working here yet, and it will included in future
commits. I splited it like that, cause the WiFi driver
will be hideusly incorporated into 6.5.3 source tree.
I wanted to include it quickly and since it originaly comes
from kernel 5.4 it requires some dirty hacks on both
kernel soruce code and the drivers sources, thus decided
to keep it as seperate, future commits.
It begs for patches for 6.5.3 codebase, though.

Defconfig 'orangepi_zero3_defconfig' has been composed by me,
based on the one in the Sunxi Linux 5.4 and other examples
I've found on the Internet.

The proper dtb has been added based on this change request on
kernel.org:
https://lore.kernel.org/linux-arm-kernel/5696551.DvuYhMxLoT@jernej-laptop/T/

This commit can be regarded as a first stable release.