Port 20U5622 WiFi chip driver to Lx 6.5.3 from 5.4
All previous features work.

The DTB has been altered, so now it has proper config
for wifi chip. The Sunxi sources, placed in previous commit
in Linux tree has been patched to work with Linux 6.5.3.

This is nowhere near to a proper port of a driver.
There are simply too many hideous hacks in the code
done by me.

But it compiles and works to some extent - scanning networks and AP mode (the latter was my goal).
Connecting to existing networks via iwconfig crashes the firmware,
but the same behaviour can be observed in images from Sunxi.

For more details see README.